Issue 3
Modifies the get_coordinate call so that is takes the name of a satellite in input.
It will be hand y to let the users choose which satellite they want to follow.

The next step is to implement an API call that will let users see all the available satellites.